+++ src/EntityGraphQL/Schema/SchemaBuilder.cs
@@ -14,72 +14,6 @@ using EntityGraphQL.Schema.FieldExtensions;
 namespace EntityGraphQL.Schema
 {
     /// <summary>
-    /// Options used by the SchemaBuilder factory with creating the SchemaProvider
-    /// </summary>
-    public class SchemaBuilderSchemaOptions
-    {
-        public static readonly Func<string, string> DefaultFieldNamer = name =>
-        {
-            return name[..1].ToLowerInvariant() + name[1..];
-        };
-        /// <summary>
-        /// Function to name fields when reading the properties from reflection. Default is camelCase
-        /// </summary>
-        public Func<string, string> FieldNamer { get; set; } = DefaultFieldNamer;
-        /// <summary>
-        /// If true (default) schema introspection will be enabled for the schema
-        /// </summary>
-        public bool IntrospectionEnabled { get; set; } = true;
-        /// <summary>
-        /// 
-        /// </summary>
-        public IGqlAuthorizationService AuthorizationService { get; set; } = new RoleBasedAuthorization();
-    }
-
-    /// <summary>
-    /// Options used by SchemaBuilder when reflection the object graph to auto create schema types & fields
-    /// </summary>
-    public class SchemaBuilderOptions
-    {
-        /// <summary>
-        /// List properties or field names to ignore. Default includes a list of EF properties
-        /// </summary>
-        public HashSet<string> IgnoreProps { get; set; } = new()
-        {
-            "Database",
-            "Model",
-            "ChangeTracker",
-            "ContextId"
-        };
-        /// <summary>
-        /// List of dotnet type names to ignore when adding types to the schema
-        /// </summary>
-        public HashSet<string> IgnoreTypes { get; set; } = new()
-        {
-            "String",
-            "Byte[]"
-        };
-        /// <summary>
-        /// If true when SchemaBuilder encounters a field that returns a list of entities and the entity has a property 
-        /// or field name Id t will also create a schema field with a singular name and an argument of id for that entity.
-        /// e.g. if it sees IEnumerable<Person> People; It will create the schema fields
-        /// {
-        ///   people: [Person]
-        ///   person(id: ID!): Person
-        /// }
-        /// </summary>
-        public bool AutoCreateFieldWithIdArguments { get; set; } = true;
-        /// <summary>
-        /// If true (default) and an enum type is encountered during reflection of the object graph it will be added to the schema as an Enum
-        /// </summary>
-        public bool AutoCreateEnumTypes { get; set; } = true;
-        /// <summary>
-        /// If true (default) and an object type is encountered during reflection of the object graph it will be added to the schema as a Type including it's fields
-        /// </summary>
-        public bool AutoCreateNewComplexTypes { get; set; } = true;
-    }
-
-    /// <summary>
     /// A simple schema provider to automatically create a query schema based on an object.
     /// Commonly used with a DbContext.
     /// </summary>
@@ -103,12 +37,29 @@ namespace EntityGraphQL.Schema
         /// <summary>
         /// Given the type TContextType recursively create a query schema based on the public properties of the object.
         /// </summary>
+        /// <param name="buildOptions">SchemaBuilderOptions to use to create the SchemaProvider and configure the rules for auto creating the schema types and fields</param>
+        /// <param name="logger">A logger to use in the schema</param>
+        /// <typeparam name="TContextType"></typeparam>
+        /// <returns></returns>
+        public static SchemaProvider<TContextType> FromObject<TContextType>(SchemaBuilderOptions? buildOptions = null, ILogger<SchemaProvider<TContextType>>? logger = null)
+        {
+            if (buildOptions == null)
+                buildOptions = new SchemaBuilderOptions();
+            var schemaOptions = new SchemaBuilderSchemaOptions();
+
+            var schema = new SchemaProvider<TContextType>(schemaOptions.AuthorizationService, schemaOptions.FieldNamer, logger, schemaOptions.IntrospectionEnabled);
+            return FromObject(schema, buildOptions);
+        }
+
+        /// <summary>
+        /// Given the type TContextType recursively create a query schema based on the public properties of the object.
+        /// </summary>
         /// <param name="schemaOptions">Options to create the SchemaProvider.</param>
         /// <param name="buildOptions">SchemaBuilderOptions to use to create the SchemaProvider and configure the rules for auto creating the schema types and fields</param>
         /// <param name="logger">A logger to use in the schema</param>
         /// <typeparam name="TContextType"></typeparam>
         /// <returns></returns>
-        public static SchemaProvider<TContextType> FromObject<TContextType>(SchemaBuilderSchemaOptions? schemaOptions = null, SchemaBuilderOptions? buildOptions = null, ILogger<SchemaProvider<TContextType>>? logger = null)
+        public static SchemaProvider<TContextType> FromObject<TContextType>(SchemaBuilderSchemaOptions? schemaOptions, SchemaBuilderOptions? buildOptions = null, ILogger<SchemaProvider<TContextType>>? logger = null)
         {
             if (buildOptions == null)
                 buildOptions = new SchemaBuilderOptions();
@@ -283,7 +234,7 @@ namespace EntityGraphQL.Schema
                     description = d.Description;
                 }
 
-                if (options.AutoCreateNewComplexTypes && (typeInfo.IsClass || typeInfo.IsInterface))
+                if ((options.AutoCreateNewComplexTypes && typeInfo.IsClass) || ((typeInfo.IsInterface || typeInfo.IsAbstract) && options.AutoCreateInterfaceTypes))
                 {
                     // add type before we recurse more that may also add the type
                     // dynamcially call generic method
@@ -301,6 +252,11 @@ namespace EntityGraphQL.Schema
 
                     var fields = GetFieldsFromObject(propType, schema, options, isInputType);
                     typeAdded.AddFields(fields);
+
+                    if (options.AutoCreateInterfaceTypes)
+                    {
+                        typeAdded.ImplementAllBaseTypes(true, true);
+                    }
                 }
                 else if (options.AutoCreateEnumTypes && typeInfo.IsEnum && !schema.HasType(propType.Name))
                 {
@@ -310,6 +266,14 @@ namespace EntityGraphQL.Schema
                 {
                     Type type = Nullable.GetUnderlyingType(propType)!;
                     schema.AddEnum(type.Name, type, description);
+                }
+                else
+                {
+                    var type = schema.GetSchemaType(propType, null);
+                    if (options.AutoCreateInterfaceTypes)
+                    {
+                        type.ImplementAllBaseTypes(true, true);
+                    }
                 }
             }
         }