+++ src/tests/EntityGraphQL.Tests/EntityQuery/GraphQLEntityQueryExtensionTests.cs
@@ -165,7 +165,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void SupportUseFilterWithOrStatement()
         {
-            var schema = SchemaBuilder.FromObject<TestDataContext>(false);
+            var schema = SchemaBuilder.FromObject<TestDataContext>(new SchemaBuilderOptions { AutoCreateFieldWithIdArguments = false });
             schema.Type<TestDataContext>().GetField("users", null)
                 .UseFilter();
             var gql = new QueryRequest
@@ -186,7 +186,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void SupportUseFilterWithAndStatement()
         {
-            var schema = SchemaBuilder.FromObject<TestDataContext>(false);
+            var schema = SchemaBuilder.FromObject<TestDataContext>(new SchemaBuilderOptions { AutoCreateFieldWithIdArguments = false });
             schema.Type<TestDataContext>().GetField("users", null)
                 .UseFilter();
             var gql = new QueryRequest
@@ -207,7 +207,7 @@ namespace EntityGraphQL.Tests
         [Fact]
         public void SupportUseFilterWithnotEqualStatement()
         {
-            var schema = SchemaBuilder.FromObject<TestDataContext>(false);
+            var schema = SchemaBuilder.FromObject<TestDataContext>(new SchemaBuilderOptions { AutoCreateFieldWithIdArguments = false });
             schema.Type<TestDataContext>().GetField("users", null)
                 .UseFilter();
             var gql = new QueryRequest