+++ src/tests/EntityGraphQL.Tests/SchemaTests/SchemaBuilderTests.cs
@@ -165,11 +165,6 @@ namespace EntityGraphQL.Tests
                     }"
             };
 
-            var context = new TestDataContext
-            {
-                Projects = new List<Project>()
-            };
-
             var res = schemaProvider.ExecuteRequest(gql, new TestSchema3(), null, null);
             Assert.Null(res.Errors);